선생님께서 쓰시는 image, ani로 교체
힘들었음

+++ GameDev/GameDev.h
@@ -5,7 +5,7 @@
 #define _USE_MATH_DEFINES
 #include <math.h>
 
-#define D2R (M_PI/180.)
+#define D2R float(M_PI/180)
 
 #include "BaseType.h"
 #include "Utility.hpp"
@@ -13,8 +13,20 @@
 #include "MainWindow.hpp"
 #include "Logger.h"
 #include "DoubleBuffer.h"
+
 #include "Control.h"
 #include "Button.h"
+#include "Window.h"
+
 #include "Sound.h"
-#include "Animation.h"
+
+#include "Object.h"
+#include "BoundaryBox.h"
+#include "statemachine.h"
+
 #include "Image.h"
+#include "Animation.h"
+
+#include "LoadAnimationResource.h"
+#include "LoadImageResource.h"
+#include "LoadSoundResource.h"
